updated auth file
