+++ lib/auth.ts
@@ -2,7 +2,7 @@ import NextAuth from "next-auth"
 import { MongoDBAdapter } from "@auth/mongodb-adapter"
 import GitHub from "next-auth/providers/github"
 import Google from "next-auth/providers/google"
-import client from "@/lib/db"
+import client from "@/lib/client"
 
 export const { handlers, auth, signIn, signOut } = NextAuth({
   adapter: MongoDBAdapter(client),
